requirements.txt
 langgraph
 langchain
 langchain-google-genai
 google-generativeai
 anthropic-mcp-client
 gitpython
 PyGithub